Introduce Gauge::Multiplexer and its demultiplexing
This allows for results to be gathered across multiple threads with
comparable accuracy to a single-threaded benchmark if throttled
afterwards such that there exists an overlap in benchmarks that keeps
(relatively) consistent CPU usage. Note that there should only be one
thread poking the GC ever, as this can lock the entire program.