Introduce Gauge.multiplex, a combination of gauged iterations
Gauge::Multiplexer can handle both demultiplexing *and* multiplexing by
behaving like a switch, like it already does. It just needs a
constructor provided gauged iterations to do multiplexing, really.
Gauge.multiplex can act as an interface for this, just it's very
annoying and inconvenient to use on its own because of how iterations
like to sink early. Bear in mind &infix:<xx> is absolutely worthless in
combination with this because it caches any sequences encountered (???).