Introduce &infix:<boot>, a warmup operator
Pre-JIT iterations may be desirable to ignore. This offers an interface
for skipping each available iterator from a Gauge::Iterator once, which
implies the instantiation of a new thread for each of these "cycles"
(until all of them have been run, at least).

This reveals that a chain of operations is processed from right to left.
&infix:<switch> is OK, just &infix:<view> needs to be right-associative
alongside &infix:<boot>.