Keep no state of Gauge in Gauge itself
The GCity of a gauged block is a property of its iterator. We erase this
information across Gauge.new unless we depend on more clone or !SET-SELF
cludging, leading to Gauge polling with the default GC behaviour despite
one being defined earlier in a chain of method calls in niche scenarios.