Introduce Gauge.pledge, a threaded mutltiplexer
This can allow any number of "covenant" threads to operate via one
iterator. A covenant replicates the results of the iteration across a
thread boundary, and is allowed to work ahead of time to reproduce a
result because its overlaid *contract* assumes this will happen. Becuase
each future iteration is predicted, the overhead of iteration can be
avoided under most circumstances. Take advantage of predictability!